requirements.txt
absl-py==1.4.0
anyio==3.6.2
apache-beam==2.46.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==3.1.0
asttokens==3.0.0
astunparse==1.6.3
attrs==21.4.0
autopep8==2.0.4
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==7.1.2
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
crcmod==1.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
docker==4.4.4
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
fastavro==1.7.3
fasteners==0.18
fastjsonschema==2.16.3
Flask==2.0.3
flatbuffers==23.3.3
fqdn==1.5.1
gast==0.4.0
google-api-core==1.32.0
google-api-python-client==1.12.11
google-apitools==0.5.31
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.17.1
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.16.2
google-cloud-bigtable==1.7.3
google-cloud-core==2.3.2
google-cloud-datastore==1.15.5
google-cloud-dlp==3.9.2
google-cloud-language==1.3.2
google-cloud-pubsub==2.13.11
google-cloud-pubsublite==1.6.0
google-cloud-recommendations-ai==0.7.1
google-cloud-resource-manager==1.6.3
google-cloud-spanner==3.26.0
google-cloud-storage==2.8.0
google-cloud-videointelligence==1.16.3
google-cloud-vision==3.1.4
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.48.2
h5py==3.8.0
hdfs==2.7.0
httplib2==0.21.0
idna==3.4
imbalanced-learn==0.12.4
importlib-metadata==6.5.0
ipykernel==6.22.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.5
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.12.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.2.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.4
jupyterlab_server==2.16.3
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.3.5
kt-legacy==1.0.5
kubernetes==12.0.1
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
ml-metadata==1.11.0
ml-pipelines-sdk==1.11.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.6.1
opt-einsum==3.3.0
orjson==3.8.10
overrides==6.5.0
packaging==20.9
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.2.0
portpicker==1.5.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.5
pure_eval==0.2.3
pyarrow==6.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydot==1.4.2
pyfarmhash==0.3.2
Pygments==2.15.1
pylint==3.1.0
pymongo==3.13.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==227
pywinpty==2.0.10
PyYAML==5.4.1
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-data-validation==1.11.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.11.0
tensorflow-model-analysis==0.42.0
tensorflow-serving-api==2.10.1
tensorflow-transform==1.11.0
termcolor==2.2.0
terminado==0.17.1
tfx==1.11.0
tfx-bsl==1.11.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.4
tornado==6.3
tqdm==4.66.2
traitlets==5.9.0
types-python-dateutil==2.9.0.20250809
typing_extensions==4.5.0
uri-template==1.2.0
uritemplate==3.0.1
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==3.6.4
wrapt==1.15.0
zipp==3.15.0
zstandard==0.21.0
